refactor: cleanup entry points
